+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   USB-Serial-JTAG controller (C3/C5/C6/C61/H2/H21/H4/P4/S3)
 - Added support for the `safe` option in `erlang:binary_to_term/2`
 - Added initial support for the OTP-compatible `atomics` module
+- Added atomics-backed support for the OTP-compatible `counters` module
 
 ### Changed
 - Updated network type db() to dbm() to reflect the actual representation of the type